Add tests for Favourites.js
Make sure that it's finding the favourites when it should,
and more importantly, not finding them when it shouldn't.

I've had a lot of problems where the favourites were being returned for
super long strings that don't match the favourite at all. I think the
letters were all there in the right order, but there were many letters
in between.